refactor: classify mouse query, dispatch layer
